+++ diive/pkgs/corrections/setto_threshold.py
@@ -1,13 +1,34 @@
 import numpy as np
 import pandas as pd
 from pandas import Series
-from pathlib import Path
 
-def setto_threshold(series: Series, threshold: float, type: str, show:bool=False,
-                    saveplot:str or Path=None) -> tuple[Series, Series]:
-    """Set values above or below a threshold value to threshold value"""
+from diive.core.utils.prints import ConsoleOutputDecorator
 
-    print(f"Set {series.name} to {type} threshold {threshold}  ...")
+
+@ConsoleOutputDecorator()
+def setto_threshold(series: Series,
+                    threshold: float,
+                    type: str,
+                    showplot: bool = False) -> Series:
+    """
+    Set values above or below a threshold value to threshold value
+
+    Args:
+        series: Data for variable that is corrected
+        threshold: Threshold value
+        type: `min` sets series values below *threshold* to *threshold*,
+            'max' sets series values above *threshold* to *threshold*
+        showplot: Show plot
+
+    Returns:
+        Corrected series
+    """
+    # caller = inspect.stack()[0].function
+    # section = HeatmapDateTime.__name__
+    # id = f"{caller}->{section}"
+    # id = setto_threshold.__name__
+    # printid = PrintID(id=id)
+    # printid.str(f"Set {series.name} to {type} threshold {threshold}  ...")
 
     outname = series.name
     series.name = "input_data"
@@ -27,7 +48,7 @@ def setto_threshold(series: Series, threshold: float, type: str, show:bool=False
     flag.loc[over_thres_ix] = 1
     flag.loc[range_ok_ix] = 0
 
-    print("QA/QC set to threshold value")
+    print(f"QA/QC set to threshold value")
     print(f"    Variable: {series.name}")
     if type == 'max':
         print(f"    Accepted → {range_ok_ix.sum()} values below max threshold of {threshold}")
@@ -39,16 +60,14 @@ def setto_threshold(series: Series, threshold: float, type: str, show:bool=False
             f"    Corrected → {over_thres_ix.sum()} values below min threshold of {threshold} were set to {threshold}")
 
     corrected_ix = flag == 1
-    series_qc = series.copy()
-    series_qc.loc[corrected_ix] = threshold
-    series_qc.rename(outname, inplace=True)
+    series_corr = series.copy()
+    series_corr.loc[corrected_ix] = threshold
+    series_corr.rename(outname, inplace=True)
 
     # Plot
-    if saveplot:
-        from diive.core.plotting.plotfuncs import quickplot_df
-        quickplot_df([series, series_qc], subplots=False,
-                     saveplot=saveplot, title=f"Set {series.name} to {type} threshold {threshold}")
+    if showplot:
+        from diive.core.plotting.plotfuncs import quickplot
+        quickplot([series, series_corr], subplots=True, showplot=showplot,
+                  title=f"Set {series.name} to {type} threshold {threshold}")
 
-    return series_qc, flag
-
-
+    return series_corr