+++ diive/pkgs/outlierdetection/missing.py
@@ -1,6 +1,9 @@
 from pandas import Series
 
+from diive.core.utils.prints import ConsoleOutputDecorator
 
+
+@ConsoleOutputDecorator()
 def missing_values(series: Series) -> Series:
     flag_name = f"QCF_MISSING_{series.name}"
     flag = Series(index=series.index, data=False)